requirements.txt
tensorflow-gpu===2.8.3
tensorflow-datasets
opencv-python
tensorboard-plugin-profile
attrs
backcall
bleach
certifi
chardet
colorama
cycler
Cython
decorator
defusedxml
entrypoints
future
gitdb2
GitPython
idna
ipykernel
ipython
ipython-genutils
ipywidgets
imageio
jedi
Jinja2
jsonschema
jupyter
jupyter-client
jupyter-console
jupyter-core
jupyterlab
jupyterlab-server
kiwisolver
MarkupSafe
matplotlib
mistune
nbconvert
nbdime
nbformat
notebook
numpy
pandocfilters
parso
pexpect
pickleshare
Pillow
prometheus-client
prompt-toolkit
ptyprocess
Pygments
pyparsing
pyrsistent
python-dateutil
pyzmq
qtconsole
requests
scipy
Send2Trash
six
smmap2
terminado
testpath
tornado
traitlets
urllib3
wcwidth
webencodings
widgetsnbextension
